Fixed Toilet index view, program runs after that added fileservice backend

+++ Toilet Clicker.Core/Domain/Toilet.cs
@@ -11,9 +11,9 @@ namespace Toilet_Clicker.Core.Domain
 	{
 		public Guid ID { get; set; }
 		public string ToiletName { get; set; }
-        public BigInteger Power { get; set; }
-		public BigInteger Speed { get; set; }
-		public BigInteger Score { get; set; }
+        public ulong Power { get; set; }
+		public ulong Speed { get; set; }
+		public ulong Score { get; set; }
 		public DateTime ToiletWasBorn { get; set; }
 
 		//db only
